Translate ALT key measure to ehrQL
The measure definition is parametrized by a codelist, such that it can
be applied to any number of codelists. The measures file is written to a
sub-directory of the `outputs` directory, because we anticipate a
downstream transformation action reading the measures file and writing a
transformed file, suitable for releasing and publishing.

Co-authored-by: Alice Wong <alice.wong@thedatalab.org>