Add single practice scenario
This adds a red line, which represents a "single practice scenario", to
the deciles chart. More specifically, it places the scenario chart above
the deciles chart, to give a [layered chart][1].

The values for the scenario chart are provided by the user, although
initial values are chosen at random and are within the range of the
measures table. To keep the number of values, and hence the number of
sliders, to a reasonable number, only values for the quarters covered by
the measures table are provided.

[1]: https://altair-viz.github.io/user_guide/compound_charts.html#layered-charts