Cache the repository
We use [`cache_resource`][1] to cache the repository, making it globally
available: that is, across all users, sessions, and reruns. Notice that
with this commit, a `Measure` is initialised once, and the instance is
then reused when the dashboard changes. Previously, a `Measure` was
initialised each time the dashboard changed.

Notice that `get_repository` allows us to maintain the separation
between the presentation (application) layer and the repository. We
don't want to see `import streamlit` in `app/measures.py`!

[1]: https://docs.streamlit.io/develop/api-reference/caching-and-state/st.cache_resource