Use Jobserver API to provide filenames
Reworked implementation of e563a29bd192e01f1b817808efa20442f55fa7cb.

This eliminates the human error in copy and pasting file URLs.
Unlike a codelist URL (which would contain the codelist name),
file URLs do not contain any hints about which measure the file is for,
so mistakes are very hard to spot.

The file URLs are now obtained programmatically from the Job Server API
and written to a separate file to `measures.yaml`.