Add dropdown to select from `Measure.charts`
Add a `charts` property to `Measure` to return the available charts for
that particular measure (currently only the "Practice deciles" chart is
available for all measures, but this will soon change for some of them).

Add a dropdown to the app so that the user can select from
`Measure.charts`.