Switch from SRO to OpenSAFELY Feedback
We switch metadata from that given by the SRO project (Service
Restoration Observatory) to that given by the OpenSAFELY Feedback
project.

It's hard to map from names to Job Server URLs, so we use the heuristic
that if the measure name contains *rate*, then the suffix of the
directory that contains the measure data will be `_tests`; and if the
measure name contains *results*, then the suffix of the directory that
contains the measure data will be either `_ref_tests` or `_mean_tests`.
We map from directories to URLs in Job Server's UI.